@@ -16,10 +16,11 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.1.3
+version: 1.1.4
 
 environment:
-  sdk: '>=3.4.4 <4.0.0'
+  sdk: ^3.6.0
+  flutter: ^3.27.4
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -28,11 +29,6 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  flutter:
-    sdk: flutter
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   carousel_slider: ^5.0.0
   cloud_firestore: ^5.4.5
   cupertino_icons: ^1.0.8
@@ -40,6 +36,8 @@ dependencies:
   equatable: ^2.0.3
   firebase_auth: ^5.5.2
   firebase_core: ^3.7.0
+  flutter:
+    sdk: flutter
   flutter_bloc: ^9.1.0
   flutter_svg: ^2.0.10+1
   freezed_annotation: ^2.4.4
@@ -55,19 +53,12 @@ dependencies:
   url_strategy: ^0.2.0
 
 dev_dependencies:
+  build_runner: ^2.4.15
+  flutter_lints: ^3.0.0
   flutter_test:
     sdk: flutter
-  build_runner: ^2.4.15
   freezed: ^2.5.8
   json_serializable: ^6.9.4
-  mocktail: ^1.0.0
-
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
-  flutter_lints: ^3.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
