@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 
 class Tag {
-  final String id;
-  final String label;
-  final Color color;
-
   const Tag({
     required this.id,
     required this.label,
     required this.color,
   });
+
+  final String id;
+  final String label;
+  final Color color;
 }
