@@ -8,17 +8,17 @@ import 'package:marketinya/core/repositories/user_repository.dart';
 import 'package:marketinya/core/services/firebase_service.dart';
 
 class AuthenticationRepository {
-  final FirebaseAuth _firebaseAuth = FirebaseService.auth;
-
-  final StreamController<Authentication> _controller = StreamController<Authentication>.broadcast();
-  late final StreamSubscription<User?> _authStateSubscription;
-  final userRepository = getIt<UserRepository>();
-
   AuthenticationRepository() {
     // Listen to repository state changes and handle them with _authStateListener.
     _authStateSubscription =
         _firebaseAuth.authStateChanges().listen(_authStateListener);
   }
+
+  final FirebaseAuth _firebaseAuth = FirebaseService.auth;
+  final StreamController<Authentication> _controller = StreamController<Authentication>.broadcast();
+  final userRepository = getIt<UserRepository>();
+
+  late final StreamSubscription<User?> _authStateSubscription;
 
   // Expose a stream of repository statuses.
   Stream<Authentication> get status => _controller.stream;
@@ -40,7 +40,7 @@ class AuthenticationRepository {
       password: password,
     );
 
-    User? user = userCredential.user;
+    final user = userCredential.user;
     if (user == null) {
       throw FirebaseAuthException(code: 'error-null-user');
     }
