@@ -1,13 +1,13 @@
 import 'dart:convert';
 
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart' hide User;
 import 'package:injectable/injectable.dart';
 import 'package:marketinya/core/config/log.dart';
 import 'package:marketinya/core/models/user.dart';
+import 'package:marketinya/core/services/firebase_service.dart';
 import 'package:marketinya/core/services/firestore_service.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart' hide User;
-import 'package:marketinya/core/services/firebase_service.dart';
 
 @injectable
 class UserRepository {
@@ -24,8 +24,10 @@ class UserRepository {
   Future<User?> getUser(String userId) async {
     try {
       final doc = await _firestore.getDocument(_users, userId);
-      if (!doc.exists) return null;
-      return User.fromJson(doc.data() as Map<String, dynamic>);
+      if (!doc.exists) {
+        return null;
+      }
+      return User.fromJson(doc.data()! as Map<String, dynamic>);
     } catch (e) {
       throw Exception('Failed to get user: $e');
     }
