@@ -1,10 +1,9 @@
-
 class BlogSectionModel {
-  final String header;
-  final String content;
-
   BlogSectionModel({
     required this.header,
     required this.content,
   });
+
+  final String header;
+  final String content;
 }
