@@ -3,9 +3,9 @@ import 'package:marketinya/website/widgets/appBar/custom_app_bar_mobile.dart';
 import 'package:marketinya/website/widgets/common/lime_contact_form_mobile.dart';
 import 'package:marketinya/website/widgets/common/subscription_form_mobile.dart';
 import 'package:marketinya/website/widgets/footer/footer_mobile.dart';
+import 'package:marketinya/website/widgets/home/mobile/business_info_section.dart';
 import 'package:marketinya/website/widgets/home/mobile/business_overview_section_mobile.dart';
 import 'package:marketinya/website/widgets/home/mobile/carousel_section_mobile.dart';
-import 'package:marketinya/website/widgets/home/mobile/business_info_section.dart';
 import 'package:marketinya/website/widgets/home/mobile/explore_services_section_mobile.dart';
 
 class HomeMobileLayout extends StatelessWidget {
@@ -27,7 +27,7 @@ class HomeMobileLayout extends StatelessWidget {
             ExploreServicesSectionMobile(),
             LimeContactFormMobile(),
             SubscriptionFormMobile(),
-            FooterMobile()
+            FooterMobile(),
           ],
         ),
       ),
