@@ -3,23 +3,24 @@ import 'package:marketinya/website/widgets/appBar/custom_app_bar_mobile.dart';
 import 'package:marketinya/website/widgets/blog/blog_cards_mobile.dart';
 import 'package:marketinya/website/widgets/common/lime_contact_form_mobile.dart';
 import 'package:marketinya/website/widgets/common/subscription_form_mobile.dart';
+import 'package:marketinya/website/widgets/footer/footer_mobile.dart';
 import 'package:marketinya/website/widgets/home/mobile/business_overview_section_mobile.dart';
 import 'package:marketinya/website/widgets/home/mobile/explore_services_section_mobile.dart';
-import 'package:marketinya/website/widgets/footer/footer_mobile.dart';
 
 class BlogMobileLayout extends StatelessWidget {
   const BlogMobileLayout({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return  Scaffold(
+    return Scaffold(
       appBar: const CustomAppBarMobile(activeTab: 'Блог'),
       body: SingleChildScrollView(
         child: Column(
           children: [
             const BusinessOverviewSectionMobile(
               text1: 'НОВОСТИ В МАРКЕТИНГА',
-              text2: 'Каĸво е необходимо да\nзнаем, за да имаме успешни\nреĸламни ĸампании.',
+              text2:
+                  'Каĸво е необходимо да\nзнаем, за да имаме успешни\nреĸламни ĸампании.',
               isBlog: true,
             ),
             const SizedBox(height: 65),
@@ -32,7 +33,7 @@ class BlogMobileLayout extends StatelessWidget {
             const ExploreServicesSectionMobile(),
             const LimeContactFormMobile(),
             const SubscriptionFormMobile(),
-            const FooterMobile()
+            const FooterMobile(),
           ],
         ),
       ),
